+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+### Added
+
+- phone number to user profile
+
 ## 1.0.0 (2020-05-27) Do Not Resuscitate
 
 - Added `/logged-in-user-information` route